🌈 Update(#98): css and layout style
